https://code.djangoproject.com/ticket/28518

Before:
```
In [2]: %timeit for x in City.objects.all()[:1000]: pass
106 ms ± 802 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
After:
```
In [2]: %timeit for x in City.objects.all()[:1000]: pass
81.8 ms ± 595 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```